pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=22.5",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
    "pyrefly>=0.46.3",
    "ruff>=0.14.10",
]
